remove shared-lib version of @devcycle/logger. move logger types to @devcycle/types. create dvcDefaultLogger functions in js and nodejs sdks
