--- lib/shared/logger/README.md
@@ -1,14 +1,0 @@
-# logger
-
-This library was generated with [Nx](https://nx.dev).
-
-
-## Building
-
-Run `nx build logger` to build the library.
-
-
-## Running unit tests
-
-Run `nx test logger` to execute the unit tests via [Jest](https://jestjs.io).
-
