--- lib/shared/logger/jest.config.ts
@@ -1,14 +1,0 @@
-module.exports = {
-  displayName: 'logger',
-  preset: '../jest.preset.ts',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    }
-  },
-  transform: {
-    '^.+\\.[tj]s$': 'ts-jest'
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../coverage/logger'
-};
