--- lib/shared/logger/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@devcycle/logger",
-  "version": "0.0.1",
-  "license": "MIT",
-  "type": "commonjs"
-}
