--- lib/shared/logger/project.json
@@ -1,37 +1,0 @@
-{
-  "root": "lib/shared/logger",
-  "sourceRoot": "lib/shared/logger/src",
-  "projectType": "library",
-  "targets": {
-    "npm-publish": {
-      "executor": "@nrwl/workspace:run-commands",
-      "options": {
-        "command": "../../../../scripts/npm-safe-publish.sh \"@devcycle/logger\"",
-        "cwd": "dist/lib/shared/logger"
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": [
-        "{options.outputFile}"
-      ],
-      "options": {
-        "lintFilePatterns": [
-          "lib/shared/logger/**/*.ts"
-        ]
-      }
-    },
-    "build": {
-      "executor": "@nrwl/js:tsc",
-      "outputs": [
-        "{options.outputPath}"
-      ],
-      "options": {
-        "outputPath": "dist/lib/shared/logger",
-        "main": "lib/shared/logger/src/index.ts",
-        "tsConfig": "lib/shared/logger/tsconfig.lib.json"
-      }
-    }
-  },
-  "tags": []
-}
