--- lib/shared/logger/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/logger'
-export * from './types/types'
