+++ sdk/js/src/logger.ts
@@ -1,7 +1,7 @@
 import { isNumber } from 'lodash'
 import {
     DVCDefaultLoggerOptions, DVCLogger
-} from '../types/types'
+} from '@devcycle/types'
 
 const prefix = '[DevCycle]: '
 export enum DVCLogLevels {
