+++ tsconfig.base.json
@@ -35,9 +35,6 @@
       "@devcycle/devcycle-react-sdk": [
         "sdk/react/src/index.ts"
       ],
-      "@devcycle/logger": [
-        "lib/shared/logger/src/index.ts"
-      ],
       "@devcycle/nodejs-server-sdk": [
         "sdk/nodejs/src/index.ts"
       ],
