+++ workspace.json
@@ -1,20 +1,19 @@
 {
   "version": 2,
   "projects": {
+    "example-js-web-elements-app": "examples/js/web-elements-app",
     "example-nodejs-js": "examples/nodejs/js",
     "example-nodejs-typescript": "examples/nodejs/typescript",
     "example-react-native-app": "examples/react/react-native-app",
     "example-react-next-js-app": "examples/react/next-js-app",
     "example-react-with-async-provider": "examples/react/with-async-provider",
     "example-react-with-provider": "examples/react/with-provider",
-    "example-js-web-elements-app": "examples/js/web-elements-app",
     "js": "sdk/js",
     "nodejs": "sdk/nodejs",
     "react": "sdk/react",
     "shared-bucketing": "lib/shared/bucketing",
     "shared-bucketing-as": "lib/shared/bucketing-assembly-script",
     "shared-bucketing-test-data": "lib/shared/bucketing-test-data",
-    "shared-logger": "lib/shared/logger",
     "shared-types": "lib/shared/types"
   }
 }
